Add Configuration to Spec
Adds the jsii configuration interfaces to jsii-spec module. This serves
as single global spec for jsii-configuration in package.json.

Move argument parsing and file/cli output to bin/jsii-config. This makes
it so the lib modules are less focused on the actual CLI interface. This
does however decrease unit test coverage which now only asserts that the
new configuration is correct and not that it was also written to the
package.json.
