+++ packages/jsii-config/bin/jsii-config.ts
@@ -1,4 +1,39 @@
-import main from '../lib';
+import * as yargs from 'yargs';
+import { promisify } from 'util';
+import { writeFile } from 'fs';
+import jsiiConfig from '../lib';
+
+const writeFilePromise = promisify(writeFile);
+/*
+ * Read package.json and prompt user for new or revised jsii config values.
+ */
+async function main() {
+  const argv = yargs
+    .command('$0 [args]', 'configure jsii compilation options in package.json')
+    .option('package-json', {
+      alias: 'p',
+      type: 'string',
+      description: 'location of module\'s package.json file',
+      default: './package.json'
+    })
+    .option('dry-run', {
+      alias: 'd',
+      type: 'boolean',
+      description: 'print output to stdout, don\'t write to package.json',
+      default: false
+    })
+    .help()
+    .argv;
+
+  const packageJsonLocation = argv.packageJson as string;
+  const output = await jsiiConfig(packageJsonLocation);
+
+  if (argv.dryRun) {
+    console.log(output);
+  } else {
+    await writeFilePromise(packageJsonLocation, JSON.stringify(output, null, 2));
+  }
+}
 
 main()
   .then(() => {
