+++ packages/jsii-spec/lib/index.ts
@@ -1,3 +1,4 @@
 export * from './assembly';
 export * from './name-tree';
 export * from './validate-assembly';
+export * from './configuration';
